// src/types/encounter.types.ts
export type DispositionType = "home" | "follow_up" | "admit";